requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
cachetools==5.5.0
google-genai==1.2.0
httpx==0.27.2
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0





